update: make the sidebar and topbar in main to be reusable see: #10
